+++ src/components/Alarm/Alarm.module.scss
