+++ src/components/buttons/index.ts
@@ -3,3 +3,5 @@ export * from './CountButton';
 export * from './MoreButton';
 export * from './OperationButton';
 export * from './CardButton';
+export * from './HeaderSigninButton';
+export * from './HeaderSignupButton';
