+++ src/components/cards/AlarmCard.module.scss
