+++ src/components/cards/EmptyAlarmCard.module.scss
