--- src/components/header/buttons/index.ts
@@ -1,2 +1,0 @@
-export * from './HeaderSigninButton';
-export * from './HeaderSignupButton';
